# openapi.json
{"openapi":"3.1.0","info":{"title":"SITS API","version":"0.1.0"},"paths":{"/commands/execute":{"post":{"tags":["Commands"],"summary":"Execute a shell command","description":"Executes a shell command based on the provided command string.","operationId":"execute_command_commands_execute_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CommandRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CommandResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/commands/history":{"get":{"tags":["Commands"],"summary":"Get Command History","description":"Retrieves a list of previously executed commands.","operationId":"get_command_history_endpoint_commands_history_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CommandHistoryResponse"}}}}}}},"/filesystem/create-directory":{"post":{"tags":["Filesystem"],"summary":"Create Directory","description":"Creates a new directory at the specified path.","operationId":"create_directory_filesystem_create_directory_post","parameters":[{"name":"path","in":"query","required":true,"schema":{"type":"string","title":"Path"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/filesystem/delete":{"delete":{"tags":["Filesystem"],"summary":"Delete File Or Directory","description":"Deletes a file or directory at the specified path.","operationId":"delete_file_or_directory_filesystem_delete_delete","parameters":[{"name":"path","in":"query","required":true,"schema":{"type":"string","title":"Path"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/filesystem/list":{"get":{"tags":["Filesystem"],"summary":"List Directory Contents","description":"Lists contents of the specified directory.","operationId":"list_directory_contents_filesystem_list_get","parameters":[{"name":"path","in":"query","required":true,"schema":{"type":"string","title":"Path"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/filesystem/read-file":{"get":{"tags":["Filesystem"],"summary":"Read File","description":"Reads contents of a specified file.","operationId":"read_file_filesystem_read_file_get","parameters":[{"name":"path","in":"query","required":true,"schema":{"type":"string","title":"Path"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/filesystem/write-file":{"post":{"tags":["Filesystem"],"summary":"Write File","description":"Writes content to a specified file.","operationId":"write_file_filesystem_write_file_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/FileWriteRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/filesystem/metadata":{"get":{"tags":["Filesystem"],"summary":"Get Metadata","description":"Retrieves metadata for a file or directory.","operationId":"get_metadata_filesystem_metadata_get","parameters":[{"name":"path","in":"query","required":true,"schema":{"type":"string","title":"Path"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/network/configuration":{"get":{"tags":["Network"],"summary":"Network Configuration","description":"Retrieves the current network configuration details.","operationId":"network_configuration_network_configuration_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/network/ping":{"post":{"tags":["Network"],"summary":"Ping Host","description":"Pings a given host and returns the result.","operationId":"ping_host_network_ping_post","parameters":[{"name":"host","in":"query","required":true,"schema":{"type":"string","title":"Host"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/processes/list":{"get":{"tags":["Processes"],"summary":"List Processes","description":"Lists active processes with details like PID, CPU, and memory usage.","operationId":"list_processes_processes_list_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/processes/kill":{"post":{"tags":["Processes"],"summary":"Kill Process","description":"Terminates a process by PID.","operationId":"kill_process_processes_kill_post","parameters":[{"name":"pid","in":"query","required":true,"schema":{"type":"integer","title":"Pid"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/":{"get":{"summary":"Read Root","operationId":"read_root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/health":{"get":{"summary":"Health Check","operationId":"health_check_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"CommandHistoryResponse":{"properties":{"history":{"items":{"type":"string"},"type":"array","title":"History","description":"List of previously executed commands."}},"type":"object","required":["history"],"title":"CommandHistoryResponse"},"CommandRequest":{"properties":{"command":{"type":"string","title":"Command","example":"ls -la /home/user"}},"type":"object","required":["command"],"title":"CommandRequest"},"CommandResponse":{"properties":{"output":{"type":"string","title":"Output","description":"Standard output from the command."},"error":{"type":"string","title":"Error","description":"Error output from the command."}},"type":"object","required":["output","error"],"title":"CommandResponse"},"FileWriteRequest":{"properties":{"path":{"type":"string","title":"Path"},"content":{"type":"string","title":"Content"}},"type":"object","required":["path","content"],"title":"FileWriteRequest"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}